pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-debug-gym"
version = "0.1.0"
description = "OpenEnv-compatible SQL debugging environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.6.0",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "PyYAML>=6.0.1",
  "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"
